Removed old, instated new SA
